move _log_tb_metrics
