+++ harness/determined/pytorch/__init__.py
@@ -22,6 +22,7 @@ from determined.pytorch._metric_utils import (
     _prepare_metrics_reducers,
     _reduce_metrics,
     _convert_metrics_to_numpy,
+    _log_tb_metrics
 )
 from determined.pytorch._experimental import PyTorchExperimentalContext
 from determined.pytorch._pytorch_context import PyTorchTrialContext
